requirements.txt
Scrapy==1.1.0
mysql-connector-python==2.1.3
MySQL-python==1.2.5
hjson==1.5.8

beautifulsoup4
Pillow
PyYAML
cssselect
lxml
nltk
requests
feedparser
tldextract
feedfinder2
jieba3k
python-dateutil
elasticsearch
pymysql
bleach
newspaper3k
readability-lxml
langdetect
configobj